Refine inserted-id handling for reminders
Agent-Logs-Url: https://github.com/Metatronius/work_in_progress/sessions/574b1f5d-0f08-4874-b3ee-f4e2bdbf2935

+++ app/src/test/java/com/example/work_in_progress/TaskViewModelTest.kt
@@ -193,13 +193,13 @@ class TaskViewModelTest {
     @Test
     fun addTask_invokesOnInsertedWithInsertedId() = runTest {
         whenever(repository.insert(any())).thenReturn(42L)
-        var insertedId: Int? = null
+        var insertedId: Long? = null
 
         viewModel.addTask(TaskParams(title = "Task")) { id ->
             insertedId = id
         }
 
-        Assert.assertEquals(42, insertedId)
+        Assert.assertEquals(42L, insertedId)
     }
 
     // allTasks LiveData
